+++ plugins/destination/gcs/client/write.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/storage"
-	"github.com/apache/arrow/go/v15/arrow"
+	"github.com/apache/arrow/go/v16/arrow"
 	"github.com/cloudquery/filetypes/v4/types"
 	"github.com/cloudquery/plugin-sdk/v4/message"
 	"github.com/google/uuid"