+++ plugins/destination/postgresql/pgarrow/to_arrow.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/apache/arrow/go/v15/arrow"
+	"github.com/apache/arrow/go/v16/arrow"
 	cqtypes "github.com/cloudquery/plugin-sdk/v4/types"
 )